Move ResponseMappingDecoder to a inner class of Feign class.
